+++ apps/frontend/src/libs/constants/constants.ts
@@ -1,2 +1,3 @@
 export { SIDEBAR_ITEMS } from "./sidebar-items.constant.js";
+export { HALF_PI, TAU } from "chart.js/helpers";
 export { PREVIOUS_INDEX_OFFSET, ZERO_INDEX } from "shared";