+++ apps/frontend/src/pages/quiz/libs/components/balance-wheel/libs/enums/percentage-config.enum.ts
@@ -1,0 +1,8 @@
+const PercentageConfig = {
+	DEFAULT_VALUE: 0,
+	INCREMENT_VALUE: 25,
+	MAX_VALUE: 100,
+	PERCENTAGE_INCREASE_INTERVAL: 1000,
+} as const;
+
+export { PercentageConfig };