+++ apps/mobile/src/screens/auth/components/sign-in-form/sign-in-form.tsx
@@ -53,7 +53,7 @@ const SignInForm: React.FC<Properties> = ({ onSubmit }) => {
 				errors={errors}
 				label="Email"
 				name="email"
-				placeholder="be@balance.com"
+				placeholder="name@example.com"
 			/>
 			<Input
 				accessoryRight={